+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## vNext
+
+### Chores
+
+- Switched from Hydra’s `release.nix` to Cicero’s `flake.nix` ([PR 3060](https://github.com/input-output-hk/daedalus/pull/3060))
+
 ## 5.1.0
 
 ### Chores